Refactor archival resolution & rendering
