+++ crates/sui-kv-rpc/src/v2/get_checkpoint.rs
@@ -1,25 +1,27 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use sui_kvstore::CheckpointData;
-use sui_kvstore::tables::checkpoints::col;
-use sui_kvstore::{BigTableClient, CHECKPOINTS_PIPELINE, KeyValueStoreReader};
+use sui_kvstore::{BigTableClient, CHECKPOINTS_PIPELINE, CheckpointData, KeyValueStoreReader};
 use sui_rpc::field::{FieldMask, FieldMaskTree, FieldMaskUtil};
-use sui_rpc::merge::Merge;
 use sui_rpc::proto::sui::rpc::v2::get_checkpoint_request::CheckpointId;
 use sui_rpc::proto::sui::rpc::v2::{Checkpoint, GetCheckpointRequest, GetCheckpointResponse};
 use sui_rpc_api::{
     CheckpointNotFoundError, ErrorReason, RpcError, proto::google::rpc::bad_request::FieldViolation,
 };
-use sui_storage::object_store::util::{build_object_store, fetch_checkpoint};
 use sui_types::digests::CheckpointDigest;
+
+use crate::bigtable_client::BigTableClient as LimitedBigTableClient;
+use crate::config::{PipelineStage, ResolvedStageConfig, StagesConfig};
+use crate::render;
+use crate::resolve;
 
 pub const READ_MASK_DEFAULT: &str = sui_rpc_api::read_mask_defaults::CHECKPOINT;
 
 pub async fn get_checkpoint(
     mut client: BigTableClient,
+    limited_client: LimitedBigTableClient,
+    stages: &StagesConfig,
     request: GetCheckpointRequest,
-    checkpoint_bucket: Option<String>,
 ) -> Result<GetCheckpointResponse, RpcError> {
     let read_mask = {
         let read_mask = request
@@ -32,7 +34,8 @@ pub async fn get_checkpoint(
         })?;
         FieldMaskTree::from(read_mask)
     };
-    let columns = checkpoint_columns(&read_mask);
+    let needs_full = resolve::needs_transactions_or_objects(&read_mask);
+    let columns = resolve::list_checkpoint_columns(&read_mask, needs_full);
     let checkpoint = match request.checkpoint_id {
         Some(CheckpointId::Digest(digest)) => {
             let digest = digest.parse::<CheckpointDigest>().map_err(|e| {
@@ -64,70 +67,47 @@ pub async fn get_checkpoint(
         }
     };
 
-    let message =
-        checkpoint_to_response(checkpoint, &read_mask, checkpoint_bucket.as_deref()).await?;
+    let message = if needs_full {
+        let transactions_stage = stages.stage(PipelineStage::Transactions);
+        let objects_stage = stages.stage(PipelineStage::Objects);
+        resolve_full_checkpoint(
+            limited_client,
+            checkpoint,
+            &read_mask,
+            transactions_stage,
+            objects_stage,
+        )
+        .await?
+    } else {
+        render::checkpoint_to_response(checkpoint, &read_mask)?
+    };
     Ok(GetCheckpointResponse::new(message))
 }
 
-/// Render a `CheckpointData` into the proto `Checkpoint`, populating fields
-/// according to `read_mask`. Shared by `get_checkpoint` and the
-/// v2alpha list-checkpoints handler.
-pub(crate) async fn checkpoint_to_response(
+/// Heavy path: resolve the checkpoint's transactions and (when requested)
+/// objects from BigTable, then render the full proto `Checkpoint`.
+async fn resolve_full_checkpoint(
+    limited_client: LimitedBigTableClient,
     checkpoint: CheckpointData,
     read_mask: &FieldMaskTree,
-    checkpoint_bucket: Option<&str>,
+    transactions_stage: ResolvedStageConfig,
+    objects_stage: ResolvedStageConfig,
 ) -> Result<Checkpoint, RpcError> {
-    let summary = checkpoint
+    let cp_seq = checkpoint
         .summary
-        .ok_or_else(|| anyhow::anyhow!("checkpoint summary missing"))?;
-    let sequence_number = summary.sequence_number;
-    let mut message = Checkpoint::default();
-    let summary: sui_sdk_types::CheckpointSummary = summary.try_into()?;
-    message.merge(&summary, read_mask);
+        .as_ref()
+        .map(|s| s.sequence_number)
+        .ok_or_else(|| RpcError::new(tonic::Code::Internal, "checkpoint summary column missing"))?;
 
-    if read_mask.contains(Checkpoint::SIGNATURE_FIELD) {
-        let signatures = checkpoint
-            .signatures
-            .ok_or_else(|| anyhow::anyhow!("checkpoint signatures missing"))?;
-        let signatures: sui_sdk_types::ValidatorAggregatedSignature = signatures.into();
-        message.merge(signatures, read_mask);
-    }
+    let (_, cp_data, txs, objects) = resolve::resolve_checkpoint(
+        limited_client,
+        read_mask,
+        transactions_stage,
+        objects_stage,
+        cp_seq,
+        checkpoint,
+    )
+    .await?;
 
-    if read_mask.contains(Checkpoint::CONTENTS_FIELD.name) {
-        let contents = checkpoint
-            .contents
-            .ok_or_else(|| anyhow::anyhow!("checkpoint contents missing"))?;
-        message.merge(
-            sui_sdk_types::CheckpointContents::try_from(contents)?,
-            read_mask,
-        );
-    }
-
-    if (read_mask.contains(Checkpoint::TRANSACTIONS_FIELD)
-        || read_mask.contains(Checkpoint::OBJECTS_FIELD))
-        && let Some(url) = checkpoint_bucket
-    {
-        let store = build_object_store(url, vec![]);
-        let checkpoint = fetch_checkpoint(&store, sequence_number).await?;
-
-        message.merge(&checkpoint, read_mask);
-    }
-
-    Ok(message)
-}
-
-/// Compute the set of BigTable columns needed for the given read mask.
-/// Always includes `s` (summary) since it provides sequence_number, digest, etc.
-/// Only includes `sg` (signatures) and `c` (contents) when needed.
-pub(crate) fn checkpoint_columns(mask: &FieldMaskTree) -> Vec<&'static str> {
-    let mut columns = vec![col::SUMMARY];
-
-    if mask.contains(Checkpoint::SIGNATURE_FIELD) {
-        columns.push(col::SIGNATURES);
-    }
-    if mask.contains(Checkpoint::CONTENTS_FIELD) {
-        columns.push(col::CONTENTS);
-    }
-
-    columns
+    render::render_full_checkpoint(cp_data, txs, objects.as_ref(), read_mask)
 }
