+++ crates/sui-kvstore/src/bigtable/client/mod.rs
@@ -886,6 +886,12 @@ impl BigTableClient {
         keys: Vec<Vec<u8>>,
         filter: Option<RowFilter>,
     ) -> Result<Vec<(Bytes, Vec<(Bytes, Bytes)>)>> {
+        // An empty key set must never reach `build_multi_get_request`: it would
+        // produce a `RowSet` with no keys/ranges and `rows_limit = 0`, which
+        // BigTable interprets as an unbounded full-table scan. Short-circuit.
+        if keys.is_empty() {
+            return Ok(vec![]);
+        }
         let request = self.build_multi_get_request(table_name, keys, filter);
         self.read_rows(request, table_name).await
     }
@@ -911,6 +917,12 @@ impl BigTableClient {
         keys: Vec<Vec<u8>>,
         filter: Option<RowFilter>,
     ) -> Result<futures::stream::BoxStream<'static, Result<(Bytes, Vec<(Bytes, Bytes)>)>>> {
+        // See `multi_get_internal`: an empty key set would be read as a
+        // full-table scan, so emit an empty stream instead of building a
+        // request.
+        if keys.is_empty() {
+            return Ok(futures::stream::empty().boxed());
+        }
         let request = self.build_multi_get_request(table_name, keys, filter);
         let stream = self.read_rows_stream(request, table_name).await?;
         Ok(stream.boxed())
